Fix init message
