+++ internal/mocks/rit_config_mock.go
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2020 ZUP IT SERVICOS EM TECNOLOGIA E INOVACAO SA
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package mocks
+
+import (
+	"github.com/stretchr/testify/mock"
+
+	"github.com/ZupIT/ritchie-cli/internal/pkg/config"
+)
+
+type RitConfigMock struct {
+	mock.Mock
+}
+
+func (r *RitConfigMock) Read() (config.Configs, error) {
+	args := r.Called()
+	return args.Get(0).(config.Configs), args.Error(1)
+}
+
+func (r *RitConfigMock) Write(configs config.Configs) error {
+	args := r.Called(configs)
+	return args.Error(0)
+}
